+++ src/frontend/components/Felleskomponenter/UtenlandsoppholdModal/spørsmål.ts
@@ -1,4 +1,3 @@
-import { IBarnMedISøknad } from '../../../typer/søknad';
 import { EUtenlandsoppholdÅrsak } from '../../../typer/utenlandsopphold';
 
 export enum UtenlandsoppholdSpørsmålId {
@@ -10,15 +9,6 @@ export enum UtenlandsoppholdSpørsmålId {
 }
 
 export const tilDatoUkjentLabelSpråkId = 'felles.vetikkenåravsluttes.spm';
-
-export const årsakFeilmeldingSpråkId = (barn?: IBarnMedISøknad): string =>
-    barn ? 'ombarnet.beskriveopphold.feilmelding' : 'modal.beskriveopphold.feilmelding';
-
-export const årsakLabelSpråkId = (barn?: IBarnMedISøknad): string =>
-    barn ? 'ombarnet.beskriveopphold.spm' : 'modal.beskriveopphold.spm';
-
-export const årsakSpråkId = (årsak: EUtenlandsoppholdÅrsak | '', barn?: IBarnMedISøknad): string =>
-    barn ? årsakSpråkIdsBarn[årsak] : årsakSpråkIdsSøker[årsak];
 
 export const årsakSpråkIdsSøker: Record<EUtenlandsoppholdÅrsak, string> = {
     [EUtenlandsoppholdÅrsak.FLYTTET_PERMANENT_TIL_NORGE]: 'modal.oppholdalternativ.flyttettilnorge',
@@ -38,9 +28,6 @@ export const årsakSpråkIdsBarn: Record<EUtenlandsoppholdÅrsak, string> = {
     [EUtenlandsoppholdÅrsak.OPPHOLDER_SEG_UTENFOR_NORGE]: 'ombarnet.oppholdalternativ.utenfornå',
 };
 
-export const landLabelSpråkId = (årsak: EUtenlandsoppholdÅrsak | '', barn?: IBarnMedISøknad) =>
-    barn ? landLabelSpråkIdsBarn[årsak] : landLabelSpråkIdsSøker[årsak];
-
 export const landLabelSpråkIdsBarn: Record<EUtenlandsoppholdÅrsak, string> = {
     [EUtenlandsoppholdÅrsak.FLYTTET_PERMANENT_TIL_NORGE]: 'ombarnet.hvilketlandflyttetfra.spm',
     [EUtenlandsoppholdÅrsak.FLYTTET_PERMANENT_FRA_NORGE]: 'ombarnet.hvilketlandflyttettil.spm',
@@ -55,76 +42,10 @@ export const landLabelSpråkIdsSøker: Record<EUtenlandsoppholdÅrsak, string> =
     [EUtenlandsoppholdÅrsak.OPPHOLDER_SEG_UTENFOR_NORGE]: 'modal.hvilketlandoppholderi.spm',
 };
 
-export const landFeilmeldingSpråkId = (
-    årsak: EUtenlandsoppholdÅrsak | '',
-    barn?: IBarnMedISøknad
-) => (barn ? landFeilmeldingSpråkIdsBarn[årsak] : landFeilmeldingSpråkIdsSøker[årsak]);
-
-export const landFeilmeldingSpråkIdsBarn: Record<EUtenlandsoppholdÅrsak, string> = {
-    [EUtenlandsoppholdÅrsak.FLYTTET_PERMANENT_TIL_NORGE]:
-        'ombarnet.hvilketlandflyttetfra.feilmelding',
-    [EUtenlandsoppholdÅrsak.FLYTTET_PERMANENT_FRA_NORGE]:
-        'ombarnet.hvilketlandflyttettil.feilmelding',
-    [EUtenlandsoppholdÅrsak.HAR_OPPHOLDT_SEG_UTENFOR_NORGE]:
-        'ombarnet.hvilketlandoppholdti.feilmelding',
-    [EUtenlandsoppholdÅrsak.OPPHOLDER_SEG_UTENFOR_NORGE]:
-        'ombarnet.hvilketlandoppholderi.feilmelding',
-};
-
 export const landFeilmeldingSpråkIdsSøker: Record<EUtenlandsoppholdÅrsak, string> = {
     [EUtenlandsoppholdÅrsak.FLYTTET_PERMANENT_TIL_NORGE]: 'modal.hvilketlandflyttetfra.feilmelding',
     [EUtenlandsoppholdÅrsak.FLYTTET_PERMANENT_FRA_NORGE]: 'modal.hvilketlandflyttettil.feilmelding',
     [EUtenlandsoppholdÅrsak.HAR_OPPHOLDT_SEG_UTENFOR_NORGE]:
         'modal.hvilketlandoppholdti.feilmelding',
     [EUtenlandsoppholdÅrsak.OPPHOLDER_SEG_UTENFOR_NORGE]: 'modal.hvilketlandoppholderi.feilmelding',
-};
-
-export const fraDatoFeilmeldingSpråkId = (
-    årsak: EUtenlandsoppholdÅrsak | '',
-    barn?: IBarnMedISøknad
-) => {
-    switch (årsak) {
-        case EUtenlandsoppholdÅrsak.FLYTTET_PERMANENT_FRA_NORGE:
-            return barn
-                ? 'ombarnet.nårflyttetfranorge.feilmelding'
-                : 'modal.nårflyttetfra.feilmelding';
-        default:
-            return 'felles.nårstartetoppholdet.feilmelding';
-    }
-};
-
-export const fraDatoLabelSpråkId = (årsak: EUtenlandsoppholdÅrsak | '', barn?: IBarnMedISøknad) => {
-    switch (årsak) {
-        case EUtenlandsoppholdÅrsak.FLYTTET_PERMANENT_FRA_NORGE:
-            return barn ? 'ombarnet.nårflyttetfranorge.spm' : 'modal.nårflyttetfra.spm';
-        default:
-            return 'felles.nårstartetoppholdet.spm';
-    }
-};
-
-export const tilDatoLabelSpråkId = (årsak: EUtenlandsoppholdÅrsak | '', barn?: IBarnMedISøknad) => {
-    switch (årsak) {
-        case EUtenlandsoppholdÅrsak.FLYTTET_PERMANENT_TIL_NORGE:
-            return barn ? 'ombarnet.nårflyttettilnorge.spm' : 'modal.nårflyttettilnorge.spm';
-        case EUtenlandsoppholdÅrsak.HAR_OPPHOLDT_SEG_UTENFOR_NORGE:
-            return 'felles.nåravsluttetoppholdet.spm';
-        default:
-            return 'felles.nåravsluttesoppholdet.spm';
-    }
-};
-
-export const tilDatoFeilmeldingSpråkId = (
-    årsak: EUtenlandsoppholdÅrsak | '',
-    barn?: IBarnMedISøknad
-) => {
-    switch (årsak) {
-        case EUtenlandsoppholdÅrsak.FLYTTET_PERMANENT_TIL_NORGE:
-            return barn
-                ? 'ombarnet.nårflyttettilnorge.feilmelding'
-                : 'modal.nårflyttettilnorge.feilmelding';
-        case EUtenlandsoppholdÅrsak.HAR_OPPHOLDT_SEG_UTENFOR_NORGE:
-            return 'felles.nåravsluttetoppholdet.feilmelding';
-        default:
-            return 'felles.nåravsluttesoppholdet.feilmelding';
-    }
 };
